Добавляет фунекионал
